Updated formatter/adapter classes for better runtime introspection, faster execution and added new functionality
ncChangeLogEntry can now retrieve the TableMap object for the related object

ncChangeLogAdapter has new proxy classes to ncChangeLog entry (get peer and get table map) and now holds a single instance of the Formatter class instead of creating a new one every time ->getFormatter() is called

A new field was added to the changelog array, "type"; It denotes the column type and is a performance optimization (up untill now we needed to go through the process of retrieving the colum type through table map... and it was done on every request to it).

A format string was added to the Formatter class that handles boolean columns.

Many updates to ncChangeLogUpdateChange:
 - Better handling of ColumnType, ColumnMap and TableMap (through the adapter)
 - Class name and Peer class name getters
 - renderFieldName() works as described in the docs now
 - getValue() now caches results - this is because we don't want to be firing events on every request to a value, or having to retrieve a related object multiple times
 - render() now uses the formatter instance attached to the adapter; Maybe we can now have multiple formatters for different classes?